fix: findAllMissionSymbolStack RankingService 변경
