+++ src/main/java/com/depromeet/domain/ranking/application/RankingService.java
@@ -1,5 +1,9 @@
 package com.depromeet.domain.ranking.application;
 
+import com.depromeet.domain.mission.application.MissionService;
+import com.depromeet.domain.mission.dao.MissionRepository;
+import com.depromeet.domain.mission.domain.Mission;
+import com.depromeet.domain.missionRecord.domain.MissionRecord;
 import com.depromeet.domain.ranking.dao.RankingRepository;
 import com.depromeet.domain.ranking.domain.Ranking;
 import com.depromeet.domain.ranking.dto.RankingDto;
@@ -17,6 +21,8 @@ import org.springframework.transaction.annotation.Transactional;
 public class RankingService {
 
     private final RankingRepository rankingRepository;
+    private final MissionRepository missionRepository;
+    private final MissionService missionService;
 
     @Transactional(readOnly = true)
     public List<RankingResponse> findAllRanking() {
@@ -30,7 +36,25 @@ public class RankingService {
         for (RankingDto rankingDto : rankingDtos) {
             Ranking ranking = Ranking.createRanking(rankingDto.symbolStack(), rankingDto.member());
             rankingRepository.updateSymbolStackAndMemberId(
-                    ranking.getMember().getId(), ranking.getSymbolStack());
+                    ranking.getSymbolStack(), ranking.getMember().getId());
         }
+    }
+
+    @Transactional(readOnly = true)
+    public List<RankingDto> findAllMissionSymbolStack() {
+        List<Mission> missions = missionRepository.findAllMissionWithRecords();
+        List<MissionRecord> completedMissionRecords =
+                missionService.findCompletedMissionRecords(missions);
+
+        return completedMissionRecords.stream()
+                .collect(Collectors.groupingBy(MissionRecord::getMember))
+                .entrySet()
+                .stream()
+                .map(
+                        entry ->
+                                RankingDto.of(
+                                        entry.getKey(),
+                                        missionService.symbolStackCalculate(entry.getValue())))
+                .collect(Collectors.toList());
     }
 }
